Reorganize filestructure of the client
