+++ client/src/components/anonymous/login.tsx
@@ -1,4 +1,4 @@
-import { Button } from './ui/button'
+import { Button } from '../ui/button'
 
 function Login() {
   function handleLogin() {
