+++ client/src/components/logged-in/profile.test.tsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom'
 import { render, screen } from '@testing-library/react'
-import Profile from './Profile'
+import Profile from './profile'
 import type { User } from '@/lib/types'
 
 describe('Profile', () => {
